:truck: rename: checkAuth 파일 이름을 authUtils로 변경
